Changes to pageTable. unsgnd long --> unsgnd int

+++ pageTable.hpp
@@ -17,16 +17,16 @@ class PageTable
         PageTable(int, int*);
         Level* rootLevel;
         int levelCount;
-        long unsigned int *maskArr;        // make sure to use new
-        int *shiftArr;       // make sure to use new
+        unsigned int *maskArr;        // make sure to use new
+        unsigned int *shiftArr;       // make sure to use new
         int *entryCountArr;
         unsigned int virtualAddressToPageNum(unsigned int, unsigned int, unsigned int);
         void pageInsert(PageTable*, unsigned int, unsigned int);
     private:
-        void fillMaskArr(unsigned long int*, int*, int);
-        void shiftMaskArr(unsigned long int*, int*, int);
-        unsigned long int reverseBits(unsigned long int);
-        void fillShiftArr(int*, int*, int);
+        void fillMaskArr(unsigned int*, int*, int);
+        void shiftMaskArr(unsigned int*, int*, int);
+        unsigned int reverseBits(unsigned int);
+        void fillShiftArr(unsigned int*, int*, int);
         void fillEntryCountArr(int*, int*, int);
         
 };
